# Improved performance of `SlicesIterator` (15%)