Some code simplification (10%) and inlining (5%).

```
bitmap slices 2^10      time:   [3.3655 us 3.3766 us 3.3875 us]                                
                        change: [-15.754% -15.343% -14.973%] (p = 0.00 < 0.05)

bitmap slices 2^20      time:   [3.3802 ms 3.3890 ms 3.3976 ms]                                
                        change: [-15.120% -14.776% -14.451%] (p = 0.00 < 0.05)

```